requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
Fabric3==1.14.post1
paramiko==2.6.0
psycopg2==2.8.3
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==3.3.1
